+++ app/src/main/java/com/mocalovak/cp/domain/model/Character.kt
@@ -14,7 +14,7 @@ data class Character(
     var maxHP:Int,
     var currentHP:Int,
     var gold:Int,
-    var armorClass:Int,
+    var armorClass:Float,
     var speed:Int,
     val languages:List<String>?,
     var maxMana:Int,
